+++ other/build-local/netrxavgd/netrxavgd.c
@@ -272,7 +272,7 @@ int main(int argc, char* const argv[])
 
 			if (delta_time_s != 0)
 				// round to nearest int
-				rx_bytes_per_s = (uintmax_t)(((rx_bytes - prev_rx_bytes) / delta_time_s) + 0.5);
+				rx_bytes_per_s = (uintmax_t)(((double)(rx_bytes - prev_rx_bytes) / delta_time_s) + 0.5);
 
 			char dest_buf[32] = {'\0'};
 			(void)snprintf(dest_buf, sizeof(dest_buf), "%ju", rx_bytes_per_s);
