+++ other/build-local/nettxavgd/nettxavgd.c
@@ -272,7 +272,7 @@ int main(int argc, char* const argv[])
 
 			if (delta_time_s != 0)
 				// round to nearest int
-				tx_bytes_per_s = (uintmax_t)(((tx_bytes - prev_tx_bytes) / delta_time_s) + 0.5);
+				tx_bytes_per_s = (uintmax_t)(((double)(tx_bytes - prev_tx_bytes) / delta_time_s) + 0.5);
 
 			char dest_buf[32] = {'\0'};
 			(void)snprintf(dest_buf, sizeof(dest_buf), "%ju", tx_bytes_per_s);
